+++ packages/front/src/bibliotheca/services/AuthService.ts
@@ -1,10 +1,16 @@
 import { auth, User as FirebaseUser } from 'firebase/app';
+import { gsuiteDomain } from 'bibliotheca/const';
 
 export class AuthService {
   constructor(private firebaseAuth: auth.Auth) {}
 
   async login(): Promise<void> {
     const provider = new auth.GoogleAuthProvider();
+
+    provider.setCustomParameters({
+      hd: gsuiteDomain,
+    });
+
     return this.firebaseAuth.signInWithRedirect(provider);
   }
